Extract stream-task preprocess and watermark manager tests into their own files.
