+++ src/runtime/mod.rs
@@ -4,6 +4,7 @@ pub mod consts;
 pub mod execution_graph;
 pub mod operators;
 pub mod stream_task;
+pub mod stream_task_preprocess;
 pub mod stream_task_actor;
 pub mod worker;
 pub mod worker_config_utils;
