+++ src/runtime/watermark/mod.rs
@@ -1,6 +1,9 @@
 pub mod confg;
 pub mod manager;
 
+#[cfg(test)]
+mod manager_test;
+
 pub use confg::{TimeHint, WatermarkAssignConfig};
 pub use manager::{advance_watermark_min, WatermarkAssignerState, WatermarkManager};
